// src/api/generated/models/TestStep.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { StepAttachment } from './StepAttachment';
export type TestStep = {
  action: string;
  expectedResult?: string;
  attachments?: Array<StepAttachment>;
};